外勤考核

+++ assess-app/src/main/java/com/copower/pmcc/assess/dal/basis/mapper/ProjectWorkStageMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface ProjectWorkStageMapper {
-    int countByExample(ProjectWorkStageExample example);
+    long countByExample(ProjectWorkStageExample example);
 
     int deleteByExample(ProjectWorkStageExample example);
